+++ CHANGELOG.md
@@ -20,11 +20,11 @@ Given a version number `MAJOR.MINOR.PATCH`, we increment the:
 - Backward compatibility in versions `0.0.z` is **not guaranteed** when `z` is increased.
 - Backward compatibility in versions `0.y.z` is **not guaranteed** when `y` is increased.
 
-## Unreleased
+## v0.5.3
 
 ### Fixed
 
-- Fixed internal mapping of `metaID -> previewID` that prevented stack previews from being created if stack ids contained uppercase letters.
+- Fixed issue that prevented stack previews from being created when using uppercase letters in stack IDs.
 
 ## v0.5.2